move: pieceInfo 도메인 piece패키지로 이동
