+++ src/main/java/janggi/domain/turn/Started.java
@@ -1,6 +1,6 @@
 package janggi.domain.turn;
 
-import janggi.domain.PieceInfo;
+import janggi.domain.piece.PieceInfo;
 import janggi.domain.Side;
 import janggi.domain.board.Board;
