+++ src/test/java/janggi/domain/policy/JumpPolicyTest.java
@@ -1,6 +1,6 @@
 package janggi.domain.policy;
 
-import janggi.domain.PieceInfo;
+import janggi.domain.piece.PieceInfo;
 import janggi.domain.board.BoardInterface;
 import janggi.domain.Position;
 import janggi.domain.Side;
